Tests

+++ src/main/java/ua/training/model/dao/mapper/UserMapper.java
@@ -12,7 +12,7 @@ public class UserMapper implements ObjectMapper<User> {
     @Override
     public User extractFromResultSet(ResultSet rs) throws SQLException {
         return User.builder()
-                .id(rs.getInt("user_id"))
+                .id(rs.getLong("user_id"))
                 .login(rs.getString("login"))
                 .role(User.Role.valueOf(rs.getString("role_name").toUpperCase()))
                 .balance(rs.getBigDecimal("balance"))
